+++ HISTORY.md
@@ -4,6 +4,7 @@
 * **New Features:**
     * new "this" reporter for introspection offering access to "script", "caller", "continuation" and "inputs", enabling programs to access their dynamic scope and runtime call-stack
     * "call" is now hyperized and accepts a list/table etc. of rings as procedure input, returns a list/table etc. of applying all arguments to each ring. Currently constrained to primitive data (text, number, Boolean) inputs, albeit in any dimension (e.g. lists of lists of data)
+    * new "extent" selector in the "(attribute) OF (object)" reporter's dropdown menu returns a list if width-height coordinates
     * new "values" selector in list properties reporter's dropdown, answers a new list containing the source list's unique values based on equality (i.e. a set)
     * new "sorted" and "shuffled" selectors in list properties reporter's dropdown
     * new "stage", "agent" and "script" selectors in "is a?" reporter's dropdown
@@ -73,6 +74,7 @@
 ### 2023-07-07
 * threads: changed order or priority for reporting comments
 * blocks: moved "comment" down one item in the dropdown menu
+* blocks, threads: new "extent" selector in the "(attribute) OF (object)" reporter's dropdown menu
 
 ### 2023-07-06
 * v9-rc1